Move chapter 13 boilerplate into discrete packages.
