+++ src/main/kotlin/chapter13/boilerplate/function/Function0.kt
@@ -1,0 +1,6 @@
+package chapter13.boilerplate.function
+
+import arrow.higherkind
+
+@higherkind
+data class Function0<out A>(internal val f: () -> A) : Function0Of<A>
