+++ src/main/kotlin/chapter13/sec3/listing.kt
@@ -1,8 +1,8 @@
 package chapter13.sec3
 
-import chapter13.IO
-import chapter13.fix
-import chapter13.io.monad.monad
+import chapter13.boilerplate.io.IO
+import chapter13.boilerplate.io.fix
+import chapter13.boilerplate.io.io.monad.monad
 import chapter13.stdout
 
 //tag::init1[]
